+++ CHANGELOG.md
@@ -1,3 +1,66 @@
+### [4.15.2](https://github.com/zopaUK/react-components/compare/v4.15.1...v4.15.2) (2020-08-28)
+
+
+### Bug Fixes
+
+* **alert:** set radius to 4px ([97a6b36](https://github.com/zopaUK/react-components/commit/97a6b3629ef8ad9bfc413b65c7f535fc8a02eb80))
+
+### [4.15.1](https://github.com/zopaUK/react-components/compare/v4.15.0...v4.15.1) (2020-08-24)
+
+
+### Bug Fixes
+
+* **inputs:** add backgroound color on disabled state ([1831e88](https://github.com/zopaUK/react-components/commit/1831e88096397a35f9a30a9054599c86207c8055))
+
+## [4.15.0](https://github.com/zopaUK/react-components/compare/v4.14.2...v4.15.0) (2020-08-24)
+
+
+### Features
+
+* **zopafooter:** add credit card link ([facf226](https://github.com/zopaUK/react-components/commit/facf2260920b3b400be6a0e7bf46fe4781e19c8d))
+* **zopafooter:** add savings link ([376c72f](https://github.com/zopaUK/react-components/commit/376c72fade282add97da46ecc0da70291074dca2))
+* **zopafooter:** update snapshot ([63520f1](https://github.com/zopaUK/react-components/commit/63520f18d06d4c7a48f6b58158f1a9a77e2edebf))
+
+### [4.14.2](https://github.com/zopaUK/react-components/compare/v4.14.1...v4.14.2) (2020-08-21)
+
+
+### Bug Fixes
+
+* **inputrange:** make background color transparent ([2f7418e](https://github.com/zopaUK/react-components/commit/2f7418e9cde59d6c3cb536da5998875d012c9b61))
+
+### [4.14.1](https://github.com/zopaUK/react-components/compare/v4.14.0...v4.14.1) (2020-08-21)
+
+
+### Bug Fixes
+
+* **inputrange:** make background color transparent ([227d99b](https://github.com/zopaUK/react-components/commit/227d99bdeaa2f3ebff7bcc51b88d24f2f1c5c194))
+
+## [4.14.0](https://github.com/zopaUK/react-components/compare/v4.13.1...v4.14.0) (2020-08-21)
+
+
+### Features
+
+* **help:** allow help line prop and email address fix ([6527bbc](https://github.com/zopaUK/react-components/commit/6527bbc78e968e0ab61f54243342372d9250c9d0))
+
+### [4.13.1](https://github.com/zopaUK/react-components/compare/v4.13.0...v4.13.1) (2020-08-20)
+
+
+### Bug Fixes
+
+* **notification:** adds export of notification and banner ([081e296](https://github.com/zopaUK/react-components/commit/081e29662281fbc311a6f025ccac138a7e8596a0))
+
+## [4.13.0](https://github.com/zopaUK/react-components/compare/v4.12.0...v4.13.0) (2020-08-18)
+
+
+### Features
+
+* **alert:** adds banner and notification components ([0ae8d29](https://github.com/zopaUK/react-components/commit/0ae8d29443516c15f148969d6d78b0fa0bc58891))
+
+
+### Bug Fixes
+
+* **notification:** fixes notification using non rounded corners ([429e7ce](https://github.com/zopaUK/react-components/commit/429e7ce4a7483cc827ce6b6b4271728201164c5d))
+
 ## [4.12.0](https://github.com/zopaUK/react-components/compare/v4.11.2...v4.12.0) (2020-08-17)